+++ Student/mp2_code/io/utils.hpp
@@ -5,13 +5,13 @@
 #include <arpa/inet.h>
 
 // Adds a short int (host order) to the buffer in network order
-int add_short(short int num, char *buffer);
+int add_short(uint16_t num, char *buffer);
 
 // Adds a long int (host order) to the buffer in network order
-int add_long(long int num, char *buffer);
+int add_long(uint32_t num, char *buffer);
 
 // Extracts a short int from network buffer and returns it in host order
-short int extract_short(const char *buffer);
+uint16_t extract_short(const char *buffer);
 
 // Extracts a long int from network buffer and returns it in host order
-long int extract_long(const char *buffer);
+uint32_t extract_long(const char *buffer);
